Reorganized the browsing portion of the UI following the 'drawer' design pattern.
All top level Activities are accessed via a slide out drawer.  The drawer can be opened by tapping the 'up' caret or swiping in from the left side of the screen.  The drawer is accessible from all views, however, the Activity stack is not currently cleared when switching to a new top level Activity.

Removed the Wheat theme temporarily as it doesn't work correctly with the drawer library.